[ultisnips] Testing it out
This is a very big change, and it isn't 1:1.

Frankly, the only reason I'm doing this is that I need recursive
evaluation in snippet output.  I much preferred neosnippet, but it is
nice being able to generate correct headers for reST for example.

neosnippet with support for things like the following:

    ${1:#:header}
    `repeat('=', len($1))`

would have me switch back in an instant.