Move .vim files to reST
This is a *huge* changeset, but very little is actually changed.

Bear in mind that it is *largely* mechanical, so isn’t quite the
enormous task it looks like it might have been.  Thank $deity we have
a *real* editor to do this with too; macros with motions are awesome.