Make Sphinx builds optional
On some systems I’m using Sphinx isn’t available, and it is nice to
stand up an environment without having to pull in Sphinx first.

An alternative *might* be to produce a standalone .vim file tree that is
easily usable, perhaps as a git branch with pre-built files produced
with a hook.