[vim-peekaboo] Move window configuration to autoload
Dynamically setting the window type causes a minor slowdown on startup
because of the ``xdotool`` execution.

Moving it to the initial execution doesn’t appear to be to worrisome,
but this may have to be reverted if the interaction on first use
suffers.  # Please enter the commit message for your changes. Lines
starting