❌ Remove shell script automatic chmod +x
As this only worked in the specific case of files ending in ``.sh`` it
missed an awful lot of useful cases, and caught a few where it wasn’t
needed.  The only way to fix it would be to switch to using ``FileType``
to match *and* scanning the file for a ``#!`` header, which is more
grief than occasionally having to manually run ``chmod``.

If I change my mind this will come back with support for other script
types though ;)