[QA] Switch to printf where it is more readable
There were a fair few examples of convoluted concatenation where the
``printf()`` version is a lot easier to read.  I’ll be honest here and
say the general reason the concatenation is so prevalent in here was
that I often build up the strings with ``:echo`` and then insert the
``":`` register when it works.