[vim-projectionist] 🐛 Don’t blindly run commands from templates
This functionality is inherently dangerous, and the initial version
should have thought about the consequences.

While asking the user on every instantiation is annoying it does feel
workable.  Should it become annoying to have to respond to the prompts
it wouldn’t be *too* much work to hack in state masking(using a checksum
over the template input for example).  It would need a lot more thought
being put in to it though.