[dein.vim] Always use dict/list in on_* setup
This is mainly to test the speed difference from dropping all the type
switching in dein.  It looks like it makes a very minor difference, but
at the same time the explicitness isn’t exactly costly either.

This may be reverted if it doesn’t work out, or if it won’t be accepted
upstream.